make max_requests per second float

+++ src/keboola/http_client/async_client.py
@@ -18,7 +18,7 @@ class AsyncHttpClient:
             timeout: Optional[float] = None,
             verify_ssl: bool = True,
             retry_status_codes: Optional[List[int]] = None,
-            max_requests_per_second: Optional[int] = None,
+            max_requests_per_second: Optional[float] = None,
             default_params: Optional[Dict[str, str]] = None,
             auth: Optional[tuple] = None,
             auth_header: Optional[Dict[str, str]] = None,
@@ -35,7 +35,7 @@ class AsyncHttpClient:
             timeout (Optional[float], optional): The request timeout in seconds. Defaults to None.
             verify_ssl (bool, optional): Enable or disable SSL verification. Defaults to True.
             retry_status_codes (Optional[List[int]], optional): List of status codes to retry on. Defaults to None.
-            max_requests_per_second (Optional[int], optional): Maximum number of requests per second. Defaults to None.
+            max_requests_per_second (Optional[float], optional): Maximum number of requests per second. Defaults to None.
             default_params (Optional[Dict[str, str]], optional): Default query parameters for each request.
             auth (Optional[tuple], optional): Authentication credentials for each request. Defaults to None.
             auth_header (Optional[Dict[str, str]], optional): Authentication header for each request. Defaults to None.
@@ -52,7 +52,7 @@ class AsyncHttpClient:
 
         self.limiter = None
         if max_requests_per_second:
-            one_reqeust_per_second_amount = float(1/int(max_requests_per_second))
+            one_reqeust_per_second_amount = float(1/max_requests_per_second)
             self.limiter = AsyncLimiter(1, one_reqeust_per_second_amount)
 
         self.default_headers = default_headers or {}
